MAT-6841 added codes section content
