+++ src/terminologySectionPanel/codesSection/codesSubSection/codeSubSection/ResultsSection.tsx
@@ -1,10 +1,13 @@
 import React from "react";
 import TerminologySection from "../../../../common/TerminologySection";
 
-export default function ResultsSection() {
+export default function ResultsSection({ showResultsTable }) {
   return (
     <div style={{ marginTop: "30px" }}>
-      <TerminologySection title="Results" />
+      <TerminologySection
+        title="Results"
+        showHeaderContent={showResultsTable}
+      />
     </div>
   );
 }
