+++ webpack.config.js
@@ -71,22 +71,6 @@ module.exports = (webpackConfigEnv, argv) => {
         { test: /\.json$/, type: "json" },
       ],
     },
-    // devServer: {
-    //   client: {
-    //     overlay: {
-    //       runtimeErrors: (error) => {
-    //         const ignoreErrors = [
-    //           "ResizeObserver loop limit exceeded",
-    //           "ResizeObserver loop completed with undelivered notifications.",
-    //         ];
-    //         if (ignoreErrors.includes(error.message)) {
-    //           return false;
-    //         }
-    //         return true;
-    //       },
-    //     },
-    //   },
-    // },
   };
 
   // node polyfills
